feat: implement user authentication endpoints including registration, login, logout, and password reset functionality
